backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --strip-extras
#
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
certifi==2024.6.2
    # via sentry-sdk
django==5.0.6
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-spectacular
django-filter==24.2
    # via -r requirements.in
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-gis
    #   djangorestframework-simplejwt
    #   drf-spectacular
djangorestframework-gis==1.0
    # via -r requirements.in
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
drf-dynamic-fields==0.4.0
    # via -r requirements.in
drf-spectacular==0.27.2
    # via -r requirements.in
gunicorn==22.0.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.22.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
packaging==24.1
    # via gunicorn
pillow==10.3.0
    # via -r requirements.in
psycopg==3.1.19
    # via -r requirements.in
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pymemcache==4.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
sentry-sdk==2.7.0
    # via -r requirements.in
sorl-thumbnail==12.10.0
    # via -r requirements.in
sqlparse==0.5.0
    # via django
typing-extensions==4.12.2
    # via psycopg
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.2
    # via sentry-sdk
